refactor: align archive adapters with the archive module
